[tl2] convert NBDcache to TL2 (WIP; compiles but untested)
